Move console reporter out of registry.cpp
