+++ meson.build
@@ -28,6 +28,7 @@ headers = files('include/snitch/snitch.hpp',
                 'include/snitch/snitch_matcher.hpp',
                 'include/snitch/snitch_registry.hpp',
                 'include/snitch/snitch_reporter_catch2_xml.hpp',
+                'include/snitch/snitch_reporter_console.hpp',
                 'include/snitch/snitch_reporter_teamcity.hpp',
                 'include/snitch/snitch_section.hpp',
                 'include/snitch/snitch_string.hpp',
@@ -38,16 +39,17 @@ headers = files('include/snitch/snitch.hpp',
 
 sources = files('src/snitch_append.cpp',
                'src/snitch_capture.cpp',
-               'src/snitch_catch2_xml.cpp',
                'src/snitch_cli.cpp',
                'src/snitch_console.cpp',
                'src/snitch_error_handling.cpp',
                'src/snitch_main.cpp',
                'src/snitch_matcher.cpp',
                'src/snitch_registry.cpp',
+               'src/snitch_reporter_catch2_xml.cpp',
+               'src/snitch_reporter_console.cpp',
+               'src/snitch_reporter_teamcity.cpp',
                'src/snitch_section.cpp',
                'src/snitch_string_utility.cpp',
-               'src/snitch_teamcity.cpp',
                'src/snitch_test_data.cpp')
 
 if get_option('unity_build')
