+++ tests/runtime_tests/section.cpp
@@ -301,7 +301,7 @@ TEST_CASE("section readme example", "[test macros]") {
     };
 
     framework.registry.print_callback  = print;
-    framework.registry.report_callback = &snitch::impl::default_reporter;
+    framework.registry.report_callback = &snitch::reporter::console::report;
 
     framework.test_case.func = []() {
         auto& reg = snitch::impl::get_current_test().reg;
